requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.2.0
python-multipart>=0.0.6
jinja2>=3.1.3
aiofiles>=23.2.1
openpyxl>=3.1.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
